pyproject.toml
[tool.poetry]
name = "django-cooco"
version = "0.0.1"
description = "Django application to manage cookie consents"
authors = ["roberfi"]
readme = "README.md"
packages = [{ include = "django_cooco" }]

[tool.mypy]
python_version = "3.10"
mypy_path = "$MYPY_CONFIG_FILE_DIR/django_cooco"
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "settings"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = [
    "DJ",  # flake8-django
    "E",   # pycodestyle error
    "EM",  # flake8-errmsg
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "PL",  # Pylint
    "RUF", # ruff-specific-rules
    "TC",  # flake8-type-checking
    "W",   # pycodestyle warning
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.0"
django-solo = "^2.4.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
ruff = "^0.9.2"
django-stubs = "^5.1.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
